rob updated POETs top statement of purpose
After multiple conversations with ATan we decided to LEAD with the idea that pangeo has value to you the scientist even if -- in fact *specifically if* you are not in a position to run massive-scale computations. The new statement of purpose also turns the corner into the tactical consideration of ***how*** to participate in terms of where data goes or who builds a pangeo instance.